+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * [ENHANCEMENT] Improve TraceQL regex performance in certain queries. [#3139](https://github.com/grafana/tempo/pull/3139) (@joe-elliott)
 * [ENHANCEMENT] Improve TraceQL performance in complex queries. [#3113](https://github.com/grafana/tempo/pull/3113) (@joe-elliott)
 * [BUGFIX] Fix compactor ignore configured S3 headers [#3149](https://github.com/grafana/tempo/pull/3154) (@Batkilin) 
+* [BUGFIX] Prevent building parquet iterators that would loop forever. [#](https://github.com/grafana/tempo/pull/) (@mapno)
 
 ## v2.3.0 / 2023-10-30